% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate.metab}
\alias{aggregate.metab}
\title{Aggregete metabolism data}
\usage{
\method{aggregate}{metab}(x, by = "weeks", na.action = na.pass,
  alpha = 0.05, ...)
}
\arguments{
\item{x}{input data object as returned by \code{\link{ecometab}}}

\item{by}{character string indicating aggregation period}

\item{na.action}{function for treating missing data, default \code{na.pass}}

\item{alpha}{level for estimating confidence intervals in aggregated data}

\item{...}{arguments passed to or from other methods}
}
\description{
Aggregate metabolism data for a metab object
}

