% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dectime.R
\name{dectime}
\alias{dectime}
\title{Create decimal time using day on a 24 hour scale}
\usage{
dectime(dat_in)
}
\arguments{
\item{dat_in}{data frame input with time vector as posix, must be preprocessed with \code{\link{met_day_fun}}}
}
\description{
Create decimal time using day on a 24 hour scale in prep for weighted regression
}

