% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.R, R/ecometab.R, R/plot.R
\name{aggregate.metab}
\alias{aggregate.metab}
\alias{ecometab}
\alias{ecometab.default}
\alias{plot.metab}
\title{Ecosystem metabolism}
\usage{
\method{aggregate}{metab}(x, by = "weeks", na.action = na.pass,
  alpha = 0.05, ...)

ecometab(dat_in, ...)

\method{ecometab}{default}(dat_in, DO_var = "DO_mgl", depth_val = NULL,
  metab_units = "mmol", bott_stat = FALSE, ...)

\method{plot}{metab}(x, by = "months", metab_units = "mmol", alpha = 0.05,
  width = 10, pretty = TRUE, ...)
}
\arguments{
\item{x}{input object to plot}

\item{by}{chr string describing aggregation period, passed to \code{\link{aggregate}}. See details for accepted values.}

\item{na.action}{function for treating missing data, default \code{na.pass}}

\item{alpha}{numeric indicating alpha level for confidence intervals in aggregated data. Use \code{NULL} to remove from the plot.}

\item{...}{arguments passed to or from other methods}

\item{dat_in}{Input data frame which must include time series of dissolved oxygen (mg L-1)}

\item{DO_var}{chr string indicating the name of the column with the dissolved oxygen variable for estimating metabolism}

\item{depth_val}{alternative value to use for station depth}

\item{metab_units}{chr indicating desired units of output for oxygen, either as mmol or grams}

\item{bott_stat}{logical if air-sea gas exchange is removed from the estimate}

\item{width}{numeric indicating width of top and bottom segments on error bars}

\item{pretty}{logical indicating use of predefined plot aesthetics}
}
\value{
Returns an aggregated metabolism \code{\link[base]{data.frame}}.

A data frame with daily integrated metabolism estimates including gross produciton (Pg), total respiration (Rt), and net ecosystem metabolism (NEM).

A \code{\link[ggplot2]{ggplot}} object which can be further modified.
}
\description{
Estimate ecosystem metabolism using the Odum open-water method.  Estimates of daily integrated gross production, total respiration, and net ecosystem metabolism are returned.
}
\details{
The aggregate method summarizes metabolism data by averaging across set periods of observation. Confidence intervals are also returned based on the specified alpha level.  It is used within the \code{\link{plot}} function to view summarized metabolism results.  Data can be aggregated by \code{'years'}, \code{'quarters'}, \code{'months'}, or \code{'weeks'} for the supplied function, which defaults to the \code{\link[base]{mean}}.

Input data include both water quality and weather time series, which are typically collected with independent instrument systems.  This requires merging of the time series datasets.

The open-water method is a common approach to quantify net ecosystem metabolism using a mass balance equation that describes the change in dissolved oxygen over time from the balance between photosynthetic and respiration processes, corrected using an empirically constrained air-sea gas diffusion model (see Ro and Hunt 2006, Thebault et al. 2008).  The diffusion-corrected DO flux estimates are averaged separately over each day and night of the time series. The nighttime average DO flux is used to estimate respiration rates, while the daytime DO flux is used to estimate net primary production. To generate daily integrated rates, respiration rates are assumed constant such that hourly night time DO flux rates are multiplied by 24. Similarly, the daytime DO flux rates are multiplied by the number of daylight hours, which varies with location and time of year, to yield net daytime primary production. Respiration rates are subtracted from daily net production estimates to yield gross production rates.  The metabolic day is considered the 24 hour period between sunsets on two adjacent calendar days.

Areal rates for gross production and total respiration are based on volumetric rates normalized to the depth of the water column at the sampling location, which is assumed to be well-mixed, such that the DO sensor is reflecting the integrated processes in the entire water column (including the benthos).  Water column depth is calculated as the mean value of the depth variable across the time series.  Depth values are floored at one meter for very shallow stations and 0.5 meters is also added to reflect the practice of placing sensors slightly off of the bottom.  Additionally, the air-sea gas exchange model is calibrated with wind data either collected at, or adjusted to, wind speed at 10 m above the surface. The metadata should be consulted for exact height.  The value can be changed manually using a \code{height} argument, which is passed to \code{\link{f_calcKL}}.

A minimum of three records are required for both day and night periods to calculate daily metabolism estimates.  Occasional missing values for air temperature, barometric pressure, and wind speed are replaced with the climatological means (hourly means by month) for the period of record using adjacent data within the same month as the missing data.

All DO calculations within the function are done using molar units (e.g., mmol O2 m-3).

The specific approach for estimating metabolism with the open-water method is described in Caffrey et al. 2013 and references cited therein.

The plotting method plots daily metabolism estimates using different aggregatin periods.  Accepted aggregation periods are \code{'years'}, \code{'quarters'}, \code{'months'}, \code{'weeks'}, and \code{'days'} (if no aggregation is preferred).

By default, \code{pretty = TRUE} will return a \code{\link[ggplot2]{ggplot}} object with predefined aesthetics.  Setting \code{pretty = FALSE} will return the plot with minimal modifications to the \code{\link[ggplot2]{ggplot}} object.  Use the latter approach for easier customization of the plot.
}
\examples{
\dontrun{
## import sample data
data(SAPDC)

# metadata for the location
tz <- 'America/Jamaica'
lat <- 31.39
long <- -89.28

# estimate ecosystem metabolism using observed DO time series
metab <- ecometab(SAPDC, DO_var = 'DO_obs', tz = tz,
 lat = lat, long = long)

## change aggregation period and alpha
aggregate(metab, by = 'months', alpha = 0.1)
}
\dontrun{
data(SAPDC)

# metadata for the location
tz <- 'America/Jamaica'
lat <- 31.39
long <- -89.28

# estimate ecosystem metabolism using observed DO time series
metab <- ecometab(SAPDC, DO_var = 'DO_obs', tz = tz,
 lat = lat, long = long)
}
\dontrun{
## import sample data
data(SAPDC)

# metadata for the location
tz <- 'America/Jamaica'
lat <- 31.39
long <- -89.28

# estimate ecosystem metabolism using observed DO time series
metab <- ecometab(SAPDC, DO_var = 'DO_obs', tz = tz,
 lat = lat, long = long)

## plot
plot(metab)

## change alpha, aggregation period, widths
plot(metab, by = 'quarters', alpha = 0.1, widths = 0)

## plot daily raw, no aesthetics
plot(metab, by = 'days', pretty = FALSE)
}
}
\references{
Caffrey JM, Murrell MC, Amacker KS, Harper J, Phipps S, Woodrey M. 2013. Seasonal and inter-annual patterns in primary production, respiration and net ecosystem metabolism in 3 estuaries in the northeast Gulf of Mexico. Estuaries and Coasts. 37(1):222-241.

Odum HT. 1956. Primary production in flowing waters. Limnology and Oceanography. 1(2):102-117.

Ro KS, Hunt PG. 2006. A new unified equation for wind-driven surficial oxygen transfer into stationary water bodies. Transactions of the American Society of Agricultural and Biological Engineers. 49(5):1615-1622.

Thebault J, Schraga TS, Cloern JE, Dunlavey EG. 2008. Primary production and carrying capacity of former salt ponds after reconnection to San Francisco Bay. Wetlands. 28(3):841-851.
}
\seealso{
\code{\link[stats]{aggregate}}, \code{\link{ecometab}}

\code{\link{f_calcKL}} for estimating the oxygen mass transfer coefficient used with the air-sea gas exchange model and \code{\link{met_day_fun}} for identifying the metabolic day for each observation in the time series

\code{\link{ecometab}}
}

