% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calcKL.R
\name{f_calcKL}
\alias{f_calcKL}
\title{Calculate oxygen mass transfer coefficient}
\usage{
f_calcKL(Temp, Sal, ATemp, WSpd, BP, Height = 10)
}
\arguments{
\item{Temp}{numeric for water temperature (C)}

\item{Sal}{numeric for salinity (ppt)}

\item{ATemp}{numeric for air temperature (C)}

\item{WSpd}{numeric for wind speed (m/s)}

\item{BP}{numeric for barometric pressure (mb)}

\item{Height}{numeric for height of anemometer (meters)}
}
\description{
Calculate oxygen mass transfer coefficient using equations in Thiebault et al. 2008.  Output is used to estimate the volumetric reaeration coefficient for ecosystem metabolism.
}
\details{
This function is used within the \code{\link{ecometab}} function and should not be used explicitly.
}
\references{
Ro KS, Hunt PG. 2006. A new unified equation for wind-driven surficial oxygen transfer into stationary water bodies. Transactions of the American Society of Agricultural and Biological Engineers. 49(5):1615-1622.

Thebault J, Schraga TS, Cloern JE, Dunlavey EG. 2008. Primary production and carrying capacity of former salt ponds after reconnection to San Francisco Bay. Wetlands. 28(3):841-851.
}
\seealso{
\code{\link{ecometab}}
}
