% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_calcWanninkhof.R
\name{f_calcWanninkhof}
\alias{f_calcWanninkhof}
\title{Calculate gas transfer velocity for Wanninkhof equation}
\usage{
f_calcWanninkhof(Temp, Sal, WSpd)
}
\arguments{
\item{Temp}{numeric for water temperature (C)}

\item{Sal}{numeric for salinity (ppt)}

\item{WSpd}{numeric for wind speed (m/s)}
}
\value{
numeric vector of Kw in m/d
}
\description{
Calculate gas transfer velocity for Wanninkhof equation
}
\details{
Output is Kw vector that is alternative to calculating KL using Thiebault et al. 2008 (\code{\link{f_calcKL}}). Interpreted as oxygen mass transfer coefficient.
}
\examples{
data(SAPDC)
f_calcWanninkhof(SAPDC$Temp, SAPDC$Sal, SAPDC$WSpd)
}
\references{
Wanninkhof, R. 2014. Relationship between wind speed and gas exchange over the ocean revisited. Limnology and Oceanograpy Methods. 12(6):351-362. 10.4319/lom.2014.12.351
}
