% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_day_fun.R
\name{met_day_fun}
\alias{met_day_fun}
\title{Identify metabolic days in a swmpr time series}
\usage{
met_day_fun(dat_in, tz, lat, long)
}
\arguments{
\item{dat_in}{data.frame}

\item{tz}{chr string for timezone, e.g., 'America/Chicago'}

\item{lat}{numeric for latitude}

\item{long}{numeric for longitude (negative west of prime meridian)}
}
\description{
Identify metabolic days in a time series based on sunrise and sunset times for a location and date.  The metabolic day is considered the 24 hour period between sunsets for two adjacent calendar days.
}
\details{
This function is only used within \code{\link{ecometab}} and should not be called explicitly.
}
\seealso{
\code{\link{ecometab}}
}

