% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxySchmidt.R
\name{oxySchmidt}
\alias{oxySchmidt}
\title{Calculate Schmidt number for oxygen}
\usage{
oxySchmidt(Temp, Sal)
}
\arguments{
\item{Temp}{numeric for water temperature (C)}

\item{Sal}{numeric for salinity (ppt)}
}
\value{
Cald Sc at given salinity for oxygen, unitless
}
\description{
Calculate Schmidt number for oxygen
}
\examples{
data(SAPDC)
oxySchmidt(SAPDC$Temp, SAPDC$Sal)
}
