% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAPDC.R
\docType{data}
\name{SAPDC}
\alias{SAPDC}
\title{Sample dataset for weighted regression, Sapelo Island Dean Creek station, 2012 data}
\format{\preformatted{'data.frame':	17568 obs. of  10 variables:
 $ DateTimeStamp: POSIXct, format: "2012-01-01 00:00:00" "2012-01-01 00:30:00" "2012-01-01 01:00:00" "2012-01-01 01:30:00" ...
 $ Temp         : num  14.9 14.9 14.9 14.8 14.7 14.7 14.6 14.6 14.4 14.2 ...
 $ Sal          : num  33.3 33.4 33.4 33.3 33.2 33.3 33.4 33.4 33.3 33.3 ...
 $ DO_obs       : num  5 5.5 5.9 6.4 6.6 6.1 5.7 5.2 4.6 4 ...
 $ Depth        : num  0.78 0.95 1.06 1.12 1.11 1.06 0.96 0.81 0.64 0.46 ...
 $ ATemp        : num  11.9 11.3 9.9 10 11.4 10.7 10.2 8.9 10.4 11 ...
 $ BP           : num  1022 1022 1021 1022 1022 ...
 $ WSpd         : num  0.5 0.6 0.6 2.4 1.3 0 0 0 0 0 ...
 $ TotPAR       : num  0 0.6 0.6 0.6 0.8 0.9 0.6 0.6 0.7 1.1 ...
 $ Tide         : num  0.891 1.001 1.073 1.111 1.125 ...
}}
\usage{
SAPDC
}
\description{
Sample dataset for weighted regression, Sapelo Island Dean Creek station, 2012 data
}
\keyword{datasets}

