% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtfun.R
\name{wtfun}
\alias{wtfun}
\title{Get weights used during weighted regression}
\usage{
wtfun(
  ref_in,
  dat_in,
  wt_vars = c("dec_time", "hour", "Tide"),
  wins = list(4, 12, NULL),
  all = FALSE,
  slice = TRUE,
  subs_only = FALSE
)
}
\arguments{
\item{ref_in}{one row of the data frame of \code{dat_in} that is teh center of the window}

\item{dat_in}{data frame for estimating weights}

\item{wt_vars}{chr string indicating names of weighting variables}

\item{wins}{numeric vecotr for windows for the three wt variables, values represent halves.  A \code{NULL} value for Tide specifies the half-window width is set automatically to one half the tidal range.}

\item{all}{logical to return all weights, rather than the product of all three}

\item{slice}{logical for subsetting \code{dat_in} for faster wt selection}

\item{subs_only}{logical for returning only wt vectors that are non-zero}
}
\description{
Get weights used during weighted regression for a single observation in the dissolved oxygen time series
}
\details{
The default behavior is to subset the data frame for faster wt selection by limiting the input the maximum window size.  Subsetted weights are recombined to equal a vector of length equal to the original data.
}
\seealso{
\code{\link{wtreg}}
}
