% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtreg_res.R
\docType{data}
\name{wtreg_res}
\alias{wtreg_res}
\title{Results from weighted regression with the \link{SAPDC} dataset}
\format{A \code{data.frame} with 17568 rows and 16 columns.
\describe{
  \item{DateTimeStamp}{POSIXct, timestamp of water quality observation}
  \item{Temp}{numeric, water temperature, celsius}
  \item{Sal}{numeric, salinity, ppt}
  \item{DO_obs}{numeric, dissolved oxygen, mg L-1}
  \item{ATemp}{numeric, air temperature, celsius}
  \item{BP}{numeric, barometric pressure, mb}
  \item{WSpd}{numeric, wind speed, m s-1}
  \item{Tide}{numeric, tide height, m, estimated from pressure data using harmonic regression}
  \item{metab_date}{Date, the metabolic day defined as the period from sunrise to sunrise on two adjacent calender days}
  \item{solar_period}{Factor, identifier that categorizes each time step as occuring during daylight (\code{sunrise}) or nighttime (\code{sunset}) based on sunrise and sunset times in the \code{value} column}
  \item{solar_time}{POSIXct, sequential sunrise and sunset times for the time series}
  \item{day_hrs}{numeric, total time of sunlight in the metabolic day}
  \item{dec_time}{numeric, decimal time in days}
  \item{hour}{numeric, hour of the day}
  \item{DO_prd}{numeric, predicted dissolved oxygen from \code{\link{wtreg}}}
  \item{DO_nrm}{numeric, detided (or normalized) dissolved oxygen from \code{\link{wtreg}}}
}}
\usage{
wtreg_res
}
\description{
Results after detiding the \link{SAPDC} sample dataset with \code{\link{wtreg}}.  The dataset is identical to SAPDC with the addition of eight columns that were used during regression.
}
\keyword{datasets}
