% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalcor.R
\name{evalcor}
\alias{evalcor}
\title{Evaluate tide and sun correlation}
\usage{
evalcor(dat_in, tz, lat, long, depth_val = "Tide", daywin = 6,
  method = "pearson", plot = TRUE, lims = c(-0.5, 0.5),
  progress = FALSE)
}
\arguments{
\item{dat_in}{Input \code{data.frame}}

\item{tz}{chr string for timezone, e.g., 'America/Chicago'}

\item{lat}{numeric for latitude}

\item{long}{numeric for longitude (negative west of prime meridian)}

\item{depth_val}{chr indicating name of the tidal height column in \code{dat_in}}

\item{daywin}{numeric for half-window width used in moving window correlation}

\item{method}{chr string for corrrelation method, passed to \code{\link[stats]{cor}}}

\item{plot}{logical to return a plot}

\item{lims}{two element numeric vector indicating y-axis limits on plot}

\item{progress}{logical if progress saved to a txt file names 'log.txt' in the working directory}
}
\value{
A \code{\link[ggplot2]{ggplot}} object if \code{plot = TRUE}, otherwise a numeric vector of the correlations for each row in the input dataset.
}
\description{
Evaluation correlation between tide change and sun angle to determine effectiveness of weighted regression
}
\details{
This function can be used before weighted regression to identify locations in the time series when tidal and solar changes are not correlated.  In general, the \code{\link{wtreg}} will be most effective when correlations between the two are zero, whereas \code{\link{wtreg}} will remove both the biological and physical components of the dissolved oxygen time series when the sun and tide are correlated.   The correlation between tide change and sun angle is estimated using a moving window for the time series, where the window width is defined by \code{daywin}.  Tide changes are estimated as angular rates for the tidal height vector and sun angles are estimated from the time of day and geographic location.

The \code{\link[foreach]{foreach}} function is used to execute the moving window correlation in parallel and will be run automatically if a backend is created.
Figure 9 in Beck et al. 2015 was created using this function.
}
\examples{
\dontrun{

data(SAPDC)

# metadata
tz <- 'America/Jamaica'
lat <- 31.39
long <- -89.28

# setup parallel backend
library(doParallel)
registerDoParallel(cores = 7)

evalcor(SAPDC, tz, lat, long, progress = TRUE)

}
}
\references{
Beck MW, Hagy III JD, Murrell MC. 2015. Improving estimates of ecosystem metabolism by reducing effects of tidal advection on dissolved oxygen time series. Limnology and Oceanography Methods. DOI: 10.1002/lom3.10062
}
\seealso{
\code{\link{wtreg}}
}

