% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteval.R
\name{meteval}
\alias{meteval}
\alias{meteval.metab}
\title{Evaluate metabolism results}
\usage{
meteval(metab_in)

\method{meteval}{metab}(metab_in, ...)
}
\arguments{
\item{metab_in}{input \code{metab} object as returned from \code{\link{ecometab}}}

\item{...}{additional arguments passed to other methods}
}
\value{
A vector of summary statistics named \code{meanPg}, \code{sdPg}, \code{anomPg}, \code{meanRt}, \code{sdRt}, \code{anomRt}, \code{DOcor}, \code{Pgcor}, and \code{Rtcor}.  See the details above for a meaning of each.
}
\description{
Evaluate metabolism results before and after weighted regression
}
\details{
This function provides summary statistics of metabolism results to evaluate the effectiveness of weighted regression.  These estimates are mean production, standard deviation of production, percent of production estimates that were anomalous, mean respiration, standard deviation of respiration, percent of respiration estimates that were anomalous, correlation of dissolved oxygen with tidal height changes, correlation of production with tidal height changes, and the correlation of respiration with tidal height changes.  The correlation estimates are based on an average of the correlations by each month in the time series from the raw data for dissolved oxygen and the daily results for the metabolic estimates.  Dissolved oxygen is correlated directly with tidal height at each time step.  The metabolic estimates are correlated with the tidal height ranges during the day for production and during the night for respiration.  Tidal height ranges are estimated from the raw data during each diurnal period for each metabolic day.

In general, useful results for weighted regression are those that remove the correlation of dissolved oxygen, production, and respiration with tidal changes.  Similarly, the mean estimates of metabolism should not change if a long time series is evaluated, whereas the standard deviation and percent anomalous estimates should decrease.

Tables 2 and 3 in Beck et al. 2015 were created using these methods.
}
\examples{
\dontrun{

# load library and sample data
# metab_obs and metab_dtd
 library(WtRegDO)
 data(metab_obs)
 data(metab_dtd)

 meteval(metab_obs)
 meteval(metab_dtd)
}
}
\references{
Beck MW, Hagy III JD, Murrell MC. 2015. Improving estimates of ecosystem metabolism by reducing effects of tidal advection on dissolved oxygen time series. Limnology and Oceanography Methods. DOI: 10.1002/lom3.10062
}
\seealso{
\code{\link{ecometab}}
}
