% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SAPDC.R
\docType{data}
\name{SAPDC}
\alias{SAPDC}
\title{Sample dataset for weighted regression}
\format{\preformatted{'data.frame':	152607 obs. of  17 variables:
 $ DateTimeStamp: POSIXct, format: "2004-04-18 16:30:00" "2004-04-18 17:00:00" "2004-04-18 17:30:00" "2004-04-18 18:00:00" ...
 $ Temp         : num  27 26.2 NA 24.7 22.8 21.3 20.6 20.3 20.1 20.1 ...
 $ Sal          : num  29.6 29.4 NA 29.3 28.7 28.2 28.5 29 29.4 29.6 ...
 $ DO_obs       : num  9 8.8 8.45 8.1 8.6 8.7 8.5 8.4 8.2 8 ...
 $ Depth        : num  0.05 0.38 NA 0.87 1.06 1.17 1.22 1.25 1.26 1.2 ...
 $ ATemp        : num  NA NA NA NA NA NA NA NA NA NA ...
 $ BP           : num  1029 1029 NA 1029 1029 ...
 $ WSpd         : num  5.3 4.8 NA 4.1 3.5 2.9 2.5 2.1 1.8 1.8 ...
 $ TotPAR       : num  710 614 NA 252 132 ...
 $ Tide         : num  0.535 0.727 0.924 1.099 1.24 ...
 $ dTide        : num  0.192 0.192 0.196 0.176 0.141 ...
 $ met.date     : Date, format: "2004-04-18" "2004-04-18" "2004-04-18" "2004-04-18" ...
 $ variable     : Factor w/ 2 levels "sunrise","sunset": 1 1 1 1 1 2 2 2 2 2 ...
 $ value        : POSIXct, format: "2004-04-18 05:53:03" "2004-04-18 05:53:03" "2004-04-18 05:53:03" "2004-04-18 05:53:03" ...
 $ day.hrs      : num  13.1 13.1 13.1 13.1 13.1 ...
 $ dec_time     : num  0.438 0.458 0.479 0.5 0.521 ...
 $ hour         : num  16.5 17 17.5 18 18.5 19 19.5 20 20.5 21 ...
}}
\usage{
SAPDC
}
\description{
Sample dataset for weighted regression
}
\keyword{datasets}

