% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtobjfun.R
\name{wtobjfun}
\alias{wtobjfun}
\title{An objective function to minimize plus weighted regression for finding optimal window widths}
\usage{
wtobjfun(
  dat_in,
  tz,
  lat,
  long,
  wins,
  strt = NULL,
  vls = c("meanPg", "sdPg", "anomPg", "meanRt", "sdRt", "anomRt"),
  parallel = F,
  progress = T
)
}
\arguments{
\item{dat_in}{input data frame}

\item{tz}{chr string specifying timezone of location, e.g., 'America/Jamaica' for EST, no daylight savings, must match the time zone in \code{dat_in$DateTimeStamp}}

\item{lat}{numeric for latitude of location}

\item{long}{numeric for longitude of location (negative west of prime meridian)}

\item{wins}{list of half-window widths to use in the order specified by \code{\link{wtfun}} (i.e., days, hours, tide height).}

\item{strt}{a \code{\link{POSIXct}} object returned by \code{\link{Sys.time}}}

\item{vls}{chr vector of summary evaluation object to optimize, see details for \code{\link{objfun}}}

\item{parallel}{logical if regression is run in parallel to reduce processing time, requires a parallel backend outside of the function}

\item{progress}{logical if progress saved to a txt file names 'log.txt' in the working directory}
}
\value{
A single numeric value to minimize, as output from \code{\link{objfun}}
}
\description{
An objective function to minimize plus weighted regression for finding optimal window widths
}
\examples{
\dontrun{

library(foreach)
library(doParallel)

data(SAPDC)

tz <- 'America/Jamaica'
lat <- 31.39
long <- -81.28

ncores <- detectCores()
cl <- makeCluster(ncores)
registerDoParallel(cl)

wtobjfun(SAPDC, tz = tz, lat = lat, long = long, strt = Sys.time(),
   wins = list(6, 6, 0.5), parallel = T)

stopCluster(cl)
}
}
\seealso{
\code{\link{objfun}}
}
